Split out genereateEnclaveProperties tests from shadowJar to ensure a 1:1 mapping with tasks
